refactor: extract auth session handlers to api/auth module
* Create api/auth/{providers,session}.rs with 4 handlers (providers, me, refresh, logout)
* Create repositories/user_repository.rs and services/auth_service.rs for user lookup
* Update lib.rs and openapi.rs to route to new module paths
