+++ src/api/auth/mod.rs
@@ -1,0 +1,5 @@
+/// Auth API handlers
+///
+/// Endpoints for OAuth providers listing, current user, token refresh, and logout.
+pub mod providers;
+pub mod session;
