+++ src/api/mod.rs
@@ -1,4 +1,5 @@
 pub mod analytics;
+pub mod auth;
 pub mod billing;
 pub mod keys;
 pub mod links;
