+++ src/repositories/mod.rs
@@ -16,5 +16,7 @@
 // Add repository modules here as they are created:
 pub mod settings_repository;
 pub mod tag_repository;
+pub mod user_repository;
 pub use settings_repository::SettingsRepository;
 pub use tag_repository::TagRepository;
+pub use user_repository::UserRepository;
